backend/pyproject.toml
[tool.poetry]
name = "nerdy-news-backend"
version = "0.1.0"
description = ""
authors = ["Prashant Nigam <62440911+Prashant4900@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "nerdy_news_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
requests = "^2.31.0"
pillow = "^9.5.0"
django-admin-list-filter-dropdown = "^1.0.3"
django-anymail = "^10.0"
django-compressor = "^4.3.1"
django-environ = "^0.10.0"
django-redis = "^5.3.0"
django-import-export = "^3.2.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.6"
redis = "^4.5.5"
sentry-sdk = "^1.25.1"
uvicorn = "^0.22.0"
whitenoise = "^6.5.0"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8-isort = "^6.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
psycopg2-binary = "^2.9.6"
pycodestyle = "^2.10.0"
pyflakes = "^3.0.1"
pylint = "^2.17.4"
pylint-django = "^2.5.3"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
